git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/FgpUtil/branches/wsfrestful@14045 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Util/src/main/java/org/gusdb/fgputil/db/pool/ConnectionPoolConfig.java
@@ -14,7 +14,7 @@ public interface ConnectionPoolConfig {
   public short getMaxActive();
   public short getMaxIdle();
   public short getMinIdle();
-  public short getMaxWait();
+  public long getMaxWait();
   
   public boolean isShowConnections();
   public long getShowConnectionsInterval();
